# [aws-rds] Make it possible to get metrics from Aurora.